remove battletags.js
